Add acousticbrainz plugin
This plugin gets the musical key and the beats per minute from
AcousticBrainz.

There is no rate limiting for AcousticBrainz, so this allows 20 calls
per second on the basis that:

a. Too many more might flood broadband routers; and
b. This is the right order of magnitude for track by track
AcousticBrainz calls to keep pace with MB album lookups which are rate
limited to 1 per second.