Update featartistsintitles.py
Made a list on which words to move from artist names to album and track titles, so it't not only "feat.". Also made an ignorelist. 
Would be better if it was in a configuration file, but as this is my first try in Python it must come in a later version :-)