Remove item_ids from the active requests queue when done
The active requests queue wasn't cleared after items were
acted upon. The result was that when the wikidata plugin
was active and the user tried to refresh the information for
a release, the message "loading album information" never
finished since new requests were just added to the current
"active" item, which wasn't really active, and so, was
never acted upon.