Add note about licensing to README.md
Picard itself is licensed under the GPLv2 which means that any plugins
written specifically for Picard might be argued to be under GPLv2
themselves due to GPLv2's viral nature, either due to being derived from
Picard to some extent, or due to linking to Picard at runtime.

To avoid any legal issues (and to make sure we have the rights to
distribute and change plugins for the future) the added sentence
declares that new plugins to the repository should be under a
GPLv2-compatible license (ie., a license "more free" than GPLv2 is also
acceptable).

Also, be declaring it as "new plugins" the statement is "grandfathering"
plugins already in the repository. Ideally these plugins should
eventually get sorted out (see also
https://github.com/metabrainz/picard-plugins/issues/5 ), but this should
at least be something we can point new contributors to so we don't make
the current license mess worse. :)