Version 0.6.5
Include ability to use multiple (concatenated) sources in tag mapping.
All artist "sources" using hidden variables (_cea_...) are now
consistently in the plural, to distinguish from standard tags. Note that
if "album" is used as a source in tag mapping, it will now be the
revised name, where composer prefixing is used. To use the original
name, use "release" as the source. Also various bug fixes, notably to
ensure that all arrangers get picked up for use in tag mapping.