v 0.6.6
Bug fixes and algorithm improvements. Allow multiple sources for a tag
(each will be appended) and use + as separator for concantenating
sources, not a comma, to distinguish from the use of comma to separate
multiples. Provide additional hidden variables ("sources") for vocalists
and instrumentalists. Option to include intermediate work levels in a
movement tag (for systems which cannot display multiple work levels).