v0.8
Major update.
Handle multiple recordings and/or multiple parents of a work. Handle
multiple albumartist composers for one track. Option to use "credited
as" name for artists. Option to exclude "solo" from instrument types.
Option to over-ride plugin options with those used when the album was
last saved. Option to keep (and append to) specified existing file tags.
Option to use (and write) SongKong-compatible work tags. Include the
work (and its parents) of which a work is an arrangement. Include
medleys in movement/part description. Allow for multiple parents of
recordings and works (and multiple parents of those). Option as to
whether to include parent works where the relationship attribute is
"part of collection". Plus bug fixes too numerous to mention!