v0.9
Additional option to clear previous file tags which are not wanted,
without interfering with cover art. Additional option to replace
instrument names with as-credited names. Also instrument names are now
saved to hidden variables tags (instruments, instruments_credited and
instruments_all) which can be mapped to file tags as required.
Sub-option added to the 'override artist options' option on the
"advanced" tab - to allow tag map details to be included or not in this
over-ride. Minor bug fixes and UI improvements.