Version 2.0.4 update
Added an option to attempt to get works and movement info from title if
there are no work relationships (requires title in form "work:
movement"). If Muso-specific genre processing is selected (or XML
reference file is provided including classical composers) and there is
no composer (because of a lack of work relationship) then the plugin
will check the listed artist against the reference list of classical
composers and, if there is a match, will populate the composer metadata
and set the genre to classical.