keep: Remove transltag
This was meant to allow users to specify tags that are not written to the file
to be kept. In the scription interfaces, one accesses those with a prefix of _;
internally they're saved with a prefix of ~. transltag was used to translate
from the latter (since we're iterating over the tags in their internal
representation) to the former. However, this is not required as tags beginning
with ~ are explicitly kept in the Metadata object a few lines later.