Fix and improve ReplayGain
1. Make it work at all on Py3.
2. Capture output to Picard debug log.

Although it runs the commands correctly, there are some issues documented in the description.

a. It won't work on file paths / names with unicode.


There is further work that can be done, however Philipp and I agreed to a major update along completely different lines, so this version is provided as an interium warts-and-all version.