Re-add tangoinfo plugin
The plugin had been removed in
6fc6f9410f8af82e38c0bec9a786d646d76b4eb5 because of a server
issue from tango.info, see [1].

This plugin has been modified to work with current versions
of Picard.

It currently needs a workaround for a server
misconfiguration for tango.info. Once this issue is
resolved on tango.info's side, the workaround can be
removed.

Versions 2.8+ of Picard will handle URL construction and the
Host header differently [2], so the workaround is guarded by
a version check.

[1]: https://github.com/metabrainz/picard-plugins/pull/172
[2]: https://github.com/metabrainz/picard/pull/2057