Add support for calculating ReplayGain on unidentified clusters
This commit allows unidentified clusters in the "left pane" to be run
through the plugin and have their ReplayGain calculated.  Previously the
plugin could only be used on identified tracks and albums in the "right
pane."

With this commit, one or more clusters can be scanned as albums, and the
album and track ReplayGain added to the files.  There is no change to
the existing / previous functionality for identified albums and tracks.