fix: lowercase ident
The variable `ident` contains a string with upper and lower case letters.
When saving in Picard, this causes everything in the resulting tag to be converted to lower case. The next time the album or track is refreshed, a difference is now detected because the plugin returns upper and lower case.
Picard compares with the existing tag, which is only written in lower case. The result is that the lower case tag is deleted and replaced with a mixed case tag.